Enabled adding multiple shops, changed ids to long

+++ app/src/main/java/com/github/lleuad0/shopsandprices/fragments/edit_product/EditProductViewModel.kt
@@ -25,7 +25,7 @@ class EditProductViewModel @Inject constructor(
 
     val stateFlow = MutableStateFlow(InfoUiState())
 
-    fun getProduct(productId: Int) {
+    fun getProduct(productId: Long) {
         getProductByIdUseCase.apply {
             this.productId = productId
         }.runOnBackground {
@@ -33,7 +33,7 @@ class EditProductViewModel @Inject constructor(
         }
     }
 
-    fun getDataForProduct(productId: Int) {
+    fun getDataForProduct(productId: Long) {
         getShopsAndPricesByProductIdUseCase.apply {
             this.productId = productId
         }.runOnBackground { stateFlow.update { state -> state.copy(shopsAndPrices = it) } }
