+++ data/src/main/java/com/github/lleuad0/shopsandprices/data/entities/ProductDb.kt
@@ -9,5 +9,5 @@ data class ProductDb(
     @ColumnInfo(name = "product_name") val productName: String,
     @ColumnInfo(name = "product_info") val productInfo: String,
 
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "product_id") val productId: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "product_id") val productId: Long = 0,
 )
