+++ domain/src/main/java/com/github/lleuad0/shopsandprices/domain/model/Product.kt
@@ -1,3 +1,3 @@
 package com.github.lleuad0.shopsandprices.domain.model
 
-data class Product(val name: String, val additionalInfo: String = "", val id: Int = 0)
+data class Product(val name: String, val additionalInfo: String = "", val id: Long = 0)
