+++ domain/src/main/java/com/github/lleuad0/shopsandprices/domain/usecase/GetShopsAndPricesByProductIdUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 
 class GetShopsAndPricesByProductIdUseCase @Inject constructor(private val localRepository: LocalRepository) :
     UseCase<List<Price>>() {
-    var productId: Int = 0
+    var productId: Long = 0
 
     override suspend fun execute(): List<Price> {
         return localRepository.getShopsAndPricesByProductId(productId)
